fix(data): return instead of decimal to serve number type at endpoint

+++ backend/civic_lantern/schemas/election_spending.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, ConfigDict
@@ -9,8 +8,8 @@ class ElectionSpending(BaseModel):
 
     cycle: int
     candidate_count: int
-    total_inside_receipts: Optional[Decimal]
-    total_inside_disbursements: Optional[Decimal]
-    total_outside_support: Optional[Decimal]
-    total_outside_oppose: Optional[Decimal]
-    global_influence_ratio: Optional[Decimal]
+    total_inside_receipts: Optional[float]
+    total_inside_disbursements: Optional[float]
+    total_outside_support: Optional[float]
+    total_outside_oppose: Optional[float]
+    global_influence_ratio: Optional[float]
